CU-3ebhr9b - Create package for compile time features
