+++ boa3/analyser/supportedstandard/standardanalyser.py
@@ -18,7 +18,7 @@ class StandardAnalyser(IAstAnalyser):
     """
 
     def __init__(self, analyser, symbol_table: Dict[str, ISymbol], log: bool = False):
-        from boa3.builtin import NeoMetadata
+        from boa3.builtin.compile_time import NeoMetadata
 
         super().__init__(analyser.ast_tree, analyser.filename, log=log)
