+++ boa3/model/builtin/compile_time/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['NeoMetadataType']
+
+from boa3.model.builtin.compile_time.neometadatatype import MetadataTypeSingleton as NeoMetadataType
